requirements.txt
# if on windows: pip install pypiwin32
autobahn[twisted]
scipy
webrtcvad
numpy
ffmpeg-python
uuid
soundfile